[ADD] add Profile, Home, Image Fragment
